feat:Cleanup warning
Description:清理代码clang everything级别的编译警告。理论上不会影响到源代码逻辑.

Log:无

Issue:清理警告

+++ widgets/colorslider.cpp
@@ -143,7 +143,7 @@ void ColorSlider::paintEvent(QPaintEvent *ev)
 
     const int offset = 3;
 
-    const qreal k = (qreal)(value() - minimum()) / (maximum() - minimum());
+    const qreal k = qreal(value() - minimum()) / qreal(maximum() - minimum());
     qreal x = ((rect.width() - 2 * offset) * k) + offset;
 
     QPen pen;
